fix bottom sheet

+++ lib/models/statistic_model.dart
@@ -5,7 +5,7 @@ class StatisticModel {
   int? active_pond;
   int? fish_live;
   int? fish_death;
-  int? fish_harvested;
+  double? fish_harvested;
   num? total_feed_dose;
   num? fishes_weight_lele;
   num? fishes_weight_nilamerah;
